% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq_example.R
\docType{data}
\name{physeq_example}
\alias{physeq_example}
\title{Example Dataset: physeq_example}
\format{
A \code{\link[phyloseq:phyloseq-class]{phyloseq-class}} object containing
simulated microbial abundance data, including an OTU table,
sample data, and a phylogenetic tree.
}
\usage{
data(Data_example)
}
\description{
A simulated phyloseq dataset used to illustrate the implementation of the \code{SpaMixed} method.
}
\examples{
data("physeq_example", package = "SpaMixed")
physeq_example
}
\keyword{datasets}
