% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_adjacency_matrix.R
\name{region_adjacency_matrix}
\alias{region_adjacency_matrix}
\title{Prepare a region-level adjacency matrix}
\usage{
region_adjacency_matrix(adj, ids = NULL, symmetrize = TRUE)
}
\arguments{
\item{adj}{A region-level neighborhood object. This can be either an
\code{spdep} \code{nb} object or a square matrix-like object representing
region adjacency.}

\item{ids}{Optional character vector of region identifiers. If provided,
\code{ids} is used as both row and column names of the adjacency matrix. The
length of \code{ids} must match the number of regions.}

\item{symmetrize}{Logical. If TRUE, the adjacency matrix is symmetrized using
\code{pmax(adj_mat, t(adj_mat))}. Default is TRUE.}
}
\value{
A numeric square region-by-region adjacency matrix with zero diagonal
and row and column names corresponding to region identifiers.
}
\description{
Converts a region-level neighborhood object into a numeric adjacency matrix
for use in spatial random effects.
}
\details{
This function prepares the region-level neighborhood matrix used to model
spatial dependence among regions or sampling locations in SpaMixed. The input
can be either an \code{spdep} \code{nb} object or a square adjacency matrix.

If \code{adj} is an \code{nb} object, the function converts it to a binary
adjacency matrix using \code{\link[spdep:nb2mat]{spdep::nb2mat}} with
\code{style = "B"}. If \code{ids} is not provided, the function attempts to use
\code{attr(adj, "region.id")} as the region identifiers.

If \code{adj} is already a matrix, it is converted to a numeric matrix. Row and
column names must be present, unless \code{ids} is provided. If \code{ids} is provided,
it is used to assign both row and column names. The diagonal is set to zero
because regions are not treated as neighbors of themselves.

When \code{symmetrize = TRUE}, the adjacency matrix is made symmetric using
\code{pmax(adj_mat, t(adj_mat))}. This is useful when the input neighborhood
object is directed or when one region lists another as a neighbor but not
vice versa. The resulting matrix can be used as the region-level graph in
spatial random-effect models, such as the \code{besagproper2} model in
\pkg{INLA}.
}
\examples{
adj <- matrix(
  c(
    0, 1, 0,
    1, 0, 1,
    0, 1, 0
  ),
  nrow = 3,
  byrow = TRUE
)

region_adj <- region_adjacency_matrix(
  adj = adj,
  ids = c("region1", "region2", "region3")
)

region_adj


}
\seealso{
\code{\link{cooccurrence_similarity}},
\code{\link{realized_niche_similarity}},
\code{\link{environmental_niche_similarity}},
\code{\link{phylogenetic_similarity}},
\code{\link{fit_SpaMixed}}
}
