\name{SpaMixed-package}
\alias{SpaMixed-package}
\alias{SpaMixed}
\docType{package}
\title{
\packageTitle{SpaMixed}
}
\description{
\packageDescription{SpaMixed}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{SpaMixed}

Maintainer: \packageMaintainer{SpaMixed}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = SpaMixed)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{SpaMixed}}
%\section{Documentation Index}{\packageIndices{SpaMixed}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
